fix: 🔧 parsing products, ingredients options

+++ src/main/fr/esgi/kiosk/models/Ingredient.java
@@ -2,20 +2,20 @@ package main.fr.esgi.kiosk.models;
 
 public class Ingredient extends  Ressource{
 
-    private int quantity;
+    private long quantity;
     double price = 0;
 
     public Ingredient(String uuid,int quantity) {
         this.uuid = uuid;
         this.quantity = quantity;
     }
-    public Ingredient(String uuid,int quantity, double price) {
+    public Ingredient(String uuid,long quantity, double price) {
         this.uuid = uuid;
         this.quantity = quantity;
         this.price = price;
     }
 
-    public int getQuantity() {
+    public long getQuantity() {
         return quantity;
     }
